Added binaries script which can be used to manage binaries.
It has two commands 'binaries generate' which generates an index of
the binaries in a particular directory.  The other is 'binaries
download' which will download the binaries listed in an index file.

By default 'binaries download' will download the latest versions of
software, but has an option to download all versions.

The requirements.txt file is updated because bin/binaries requires
argparse.  The argparse library is only needed if you plan to use this script.